refactor: separate connector refresh types
