+++ turbo/packages/connectors/src/auth-providers/types.ts
@@ -6,20 +6,23 @@ import type {
   ConnectorAuthMethodIdsByRevokeKind,
   ConnectorDeviceAuthGrantAuthMethodId,
   ConnectorAuthProviderType,
+  ConnectorRefreshInputValues,
+  ConnectorRefreshOutputValues,
   ConnectorType,
   DeviceAuthGrantConnectorType,
   RefreshTokenAccessConnectorType,
   TokenRevokeConnectorType,
 } from "../connectors";
-import type { StaticConnectorAuthClient } from "../connector-utils";
+import type {
+  ConnectorAuthClientForMethod,
+  StaticConnectorAuthClient,
+} from "../connector-utils";
 import type {
   AuthUrlResult,
   ConnectorAuthCodeAuthorizeArgs,
   ConnectorDeviceAuthorizationPollArgs,
   ConnectorDeviceAuthorizationStartArgs,
   ConnectorAuthCodeExchangeArgs,
-  ConnectorAuthProviderRefreshArgs,
-  ConnectorAuthProviderRefreshResult,
   ConnectorAuthProviderRevokeArgs,
   OAuthDeviceAuthPollResult,
   OAuthDeviceAuthStartResult,
@@ -61,6 +64,29 @@ export interface DeviceAuthGrantProvider<
 
 export interface NoneAccessProvider {
   readonly kind: "none";
+}
+
+export type ConnectorAuthProviderRefreshArgs<
+  T extends RefreshTokenAccessConnectorType,
+  Method extends ConnectorAuthMethodIdsByAccessKind<T, "refresh-token"> =
+    ConnectorAuthMethodIdsByAccessKind<T, "refresh-token">,
+> = {
+  readonly authClient: ConnectorAuthClientForMethod<T, Method>;
+  readonly inputs: ConnectorRefreshInputValues<T, Method>;
+  readonly signal: AbortSignal;
+};
+
+export interface ConnectorAuthProviderRefreshResultBase {
+  readonly outputs: Readonly<Record<string, string | undefined>>;
+  readonly expiresIn?: number;
+}
+
+export interface ConnectorAuthProviderRefreshResult<
+  T extends RefreshTokenAccessConnectorType,
+  Method extends ConnectorAuthMethodIdsByAccessKind<T, "refresh-token"> =
+    ConnectorAuthMethodIdsByAccessKind<T, "refresh-token">,
+> extends ConnectorAuthProviderRefreshResultBase {
+  readonly outputs: ConnectorRefreshOutputValues<T, Method>;
 }
 
 export interface RefreshTokenAccessProvider<
